@@ -4,31 +4,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@craco/craco": "^5.6.4",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "@types/jest": "^24.9.1",
-    "@types/node": "^12.12.55",
-    "@types/react": "^16.9.49",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-leaflet": "^2.5.2",
-    "@types/react-router-dom": "^5.1.5",
-    "craco-plugin-scoped-css": "^1.1.0",
-    "leaflet": "^1.7.1",
-    "node-sass": "^4.14.1",
-    "react": "^16.13.1",
-    "react-debounce-input": "^3.2.2",
-    "react-dom": "^16.13.1",
-    "react-leaflet": "^2.7.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
-    "typescript": "^3.7.5"
+    "@types/node": "^20.2.3",
+    "leaflet": "^1.9.4",
+    "react": "^18.2.0",
+    "react-datepicker": "^4.11.0",
+    "react-debounce-input": "^3.3.0",
+    "react-dom": "^18.2.0",
+    "react-leaflet": "^4.2.1",
+    "react-markdown": "^8.0.7",
+    "react-popper-tooltip": "^4.4.2",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.11.2",
+    "react-scripts": "^5.0.1",
+    "redux-saga": "^1.2.3",
+    "typescript": "^4.9.5"
   },
   "scripts": {
-    "start": "craco start",
-    "build": "craco build",
-    "test": "craco test",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build"
   },
@@ -48,6 +43,10 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^3.1.0"
+    "@types/leaflet": "^1.9.3",
+    "@types/react": "^18.2.7",
+    "@types/react-dom": "^18.2.4",
+    "gh-pages": "^5.0.0",
+    "sass": "^1.62.1"
   }
 }
