@@ -1,11 +1,12 @@
+import IParent from 'IParent';
 import React from 'react';
-import "./Layout.scoped.scss";
+import styles from "./Layout.module.scss";
 import MainMenu from './MainMenu';
 
-const Layout: React.FC = ({ children }) =>
-  <div className="site">
+const Layout: React.FC<IParent> = ({ children }) =>
+  <div className={styles.site}>
     <MainMenu />
-    <div className="page-content">
+    <div className={styles["page-content"]}>
       {children}
     </div>
   </div>;
